+++ example.py
@@ -10,6 +10,6 @@ import sys
 appkey = sys.argv[1]
 secrets_yaml = "empire_service_secrets.yaml"
 
-empire = Empire(appkey=appkey, secrets_yaml=secrets_yaml)
+empire = Empire(appkey=appkey, enduser="enduser handle", secrets_yaml=secrets_yaml)
 
 empire.walkthrough()